Fix global step update when the epoch is skipped (#7677)
* Fix global step update when the epoch is skipped

* Update CHANGELOG

* Move test
